refactor cli commands (DRY)
