+++ src/knowlang/cli/commands/chat.py
@@ -1,20 +1,12 @@
-"""Command implementation for the chat interface."""
 from knowlang.chat_bot import create_chatbot
 from knowlang.cli.types import ChatCommandArgs
-from knowlang.configs import AppConfig
+from knowlang.cli.utils import create_config
 from knowlang.utils import FancyLogger
 from knowlang.vector_stores import VectorStoreError
 from knowlang.vector_stores.factory import VectorStoreFactory
 
 LOG = FancyLogger(__name__)
 
-def create_config(args: ChatCommandArgs) -> AppConfig:
-    """Create configuration from file or defaults."""
-    if args.config:
-        with open(args.config, 'r') as file:
-            config_data = file.read()
-            return AppConfig.model_validate_json(config_data)
-    return AppConfig()
 
 async def chat_command(args: ChatCommandArgs) -> None:
     """Execute the chat command.
@@ -22,7 +14,7 @@ async def chat_command(args: ChatCommandArgs) -> None:
     Args:
         args: Typed command line arguments
     """
-    config = create_config(args)
+    config = create_config(args.config)
     
     # Initialize vector store
     try:
