+++ src/knowlang/cli/commands/serve.py
@@ -1,20 +1,12 @@
-"""Command implementation for the API server interface."""
 import uvicorn
 from knowlang.cli.types import ServeCommandArgs
-from knowlang.configs import AppConfig
+from knowlang.cli.utils import create_config
 from knowlang.utils import FancyLogger
 from knowlang.vector_stores.factory import VectorStoreFactory
 from knowlang.vector_stores import VectorStoreError
 
 LOG = FancyLogger(__name__)
 
-def create_config(args: ServeCommandArgs) -> AppConfig:
-    """Create configuration from file or defaults."""
-    if args.config:
-        with open(args.config, 'r') as file:
-            config_data = file.read()
-            return AppConfig.model_validate_json(config_data)
-    return AppConfig()
 
 async def serve_command(args: ServeCommandArgs) -> None:
     """Execute the serve command.
@@ -22,7 +14,7 @@ async def serve_command(args: ServeCommandArgs) -> None:
     Args:
         args: Typed command line arguments
     """
-    config = create_config(args)
+    config = create_config(args.config)
     
     # Initialize vector store
     try:
